refactor(git): extract commit-walk + diff helpers into git/commit.rs
Splits the commit/tree/line-diff helpers and the typed commit-walk surface
(all_commit_shas, new_commit_shas, commit_record, is_ancestor, has_commit) out
of git/mod.rs into a new git/commit.rs, keeping both files under the 1000-line
cap. The typed surface gives the git-history index builder a parity-preserving
view of history without touching gix directly.
